- fix moving the window around, which my last commit somehow broke. - mad speedup to the 'Regenerating the Music Catalog Database..' stage of a   music search.
After a music search, when it sits supposedly regenerating the database, it's
really not -- the regeneration stuff gets done quickly and the damn thing's
just processing all the MusicCatalogTrackAdded messages.  Solution was to add
a pair of messages to tell the UI to ignore Add messages for a time, then
regenerate the GUI tree all at once.  Instead of taking forever, it's done in
a second or two =)

+++ ui/freeamp/src/Headlines.cpp
@@ -144,7 +144,7 @@ void Headlines::WorkerThread(void)
             m_iIndex = m_iIndex % m_oHeadlines.size();
             m_pContext->target->AcceptEvent(new HeadlineMessageEvent(
                     m_oHeadlines[m_iIndex].c_str(),
-                    (m_iIndex >= m_oHeadlineURLs.size()) ? "" : 
+                    (m_iIndex >= (int)m_oHeadlineURLs.size()) ? "" : 
                      m_oHeadlineURLs[m_iIndex].c_str()));
                                        
             m_iIndex++;
